Change USB format from FAT32 to HFS+
I was NOT able to install legacy OS X (Mountain Lion) with online/recovery installer on a FAT32 USB stick.  On 2 separate attempts, I only had success with with an HFS+ formatted drive.  Formatting the drive as FAT32 will cause "Install OS X Mountain Lion" to do not show up in the OpenCore picker.  As soon as I formatted as HFS+ and repeated the steps, the installer was accessible and installed with no issues.

If FAT32 is indeed incorrect, this should be corrected, along with the screenshot.